DeliverPhone Class
String variable names improved and changed

Phone Class
Float changed to double for price

PhoneSpec Class.
Got rid of extend Customer. Had no reference in it in the class.

+++ src/liamDunne/Phone.java
@@ -27,7 +27,7 @@ public  class Phone extends Customer {
             return price;
         }
 
-        public void setPrice(final float newPrice) {
+        public void setPrice(final double newPrice) {
             this.price = newPrice;
         }
